ARM: mvebu: prepare coherency code to support more SOCs
The code that handles the coherency fabric of Armada 370 and Armada XP
in arch/arm/mach-mvebu/coherency.c made the assumption that there was
only one type of coherency fabric. Unfortunately, it turns out that
upcoming SoCs have a slightly different coherency unit.

In preparation to the introduction of the coherency support for more
SoCs, this commit:

 * Adds a new "marvell,armada-370-coherency-fabric" compatible string,
   to indicate the Armada 370/XP variant of the coherency fabric. The
   existing compatible string "marvel,coherency-fabric" is preserved
   for compatibility reasons, and has the same behavior as
   the new "marvell,armada-370-coherency-fabric".

 * Introduces a data associated to the compatible string in the
   compatible string match table, so that the code can differantiate
   the variant of coherency unit being used.

 * Separates the coherency unit initialization code into its own
   function.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>