ARM: mvebu: implement Armada 375 coherency workaround
The early revisions of Armada 375 SOCs (Z1 stepping) have a bug in the
I/O coherency unit that prevents using the normal method for the I/O
coherency barrier. The recommended workaround is to use a XOR memset
transfer to act as the I/O coherency barrier.

This involves "borrowing" a XOR engine, which gets disabled in the
Device Tree so the normal XOR driver doesn't use it. Note also that
the mvebu-mbus driver must be initialized prior to the coherency unit,
because the DRAM windows of the XOR engine must be configured during
the coherency unit initialization, due to this workaround.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>