ARM: mm: add support for HW coherent systems in PL310
When a PL310 cache with a revision >= r3p2 is used on a system that
provides hardware coherency, the outer cache sync operation is
useless, and can be skipped. Moreover, on some systems, it is harmful
as it causes deadlocks between the Marvell coherency mechanism, the
Marvell PCIe controller and the Cortex-A9.

This commit adds a new l2x0_of_init_coherent() varaitn of
l2x0_of_init(), which allows the caller to tell the L2 cache
initialization that the system has hardware coherency support enabled,
and that therefore the outer cache sync operation can be skipped if
possible.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>