Attempt at fixing 'splice optimization' when using sendfile()
When using splice to receive a file through a socket, a possible
implementation is to call splice() until there's nothing more to
to transfer (i.e. until we get a broken pipe?).

In such case, kernel_recvmsg() reports an error (or incomplete transfer)
and the correct way of completing the write to the filesystem is by
calling write_end() with copied=0. This should unlock the pages,
but avoid writing any of them to the file (?).

Not sure if any of the above makes any sense, but it fixed my
test case :-)

Signed-off-by: Ezequiel Garcia <ezequiel.garcia@free-electrons.com>