bus: mvebu: add support for Armada 380/385 in mvebu-mbus
The mvebu-mbus driver handles the special MBus mechanism of Marvell
EBU SoCs. The two new members of the EBU family, the Armada 380 and
Armada 385, are identical in terms of MBus handling, and share the
same number of windows and register organization as Armada 370/XP.

Therefore, this commit adds a new "marvell,armada380-mbus" compatible
string, whose behavior is for now identical as the existing
"marvell,armadaxp-mbus" compatible string.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>