usb: host: xhci-plat: Allow to register glue code using the device tree
This patch allow to register specific glue code for xhci controller.
It creates a structure called xhci_plat_ops to register functions
specific to a SoC. Currently there are only probe() and remove() but
it can be extended later, it was the point to create such a structure.

Each compatible string can then be associated to an instance of this
structure. In the non device tree case a default structure is used.

Signed-off-by: Gregory CLEMENT <gregory.clement@free-electrons.com>