ARM: mvebu: Enable NAND controller in A370 Reference Design board
Marvell's Armada 370 Reference Design has a NAND flash, so enable it in
the devicetree and add the partitions as prepared in the factory images.

In order to skip the driver's custom device detection and use only ONFI
detection, the marvell,keep-config parameter is used. This is needed
because we have no support for setting the timings parameters yet.

Change-Id: I07c3662bba0209c3eab85ff186e4f33c63151802
Signed-off-by: Ezequiel Garcia <ezequiel.garcia@free-electrons.com>
Signed-off-by: Jason Cooper <jason@lakedaemon.net>
Signed-off-by: Nadav Haklai <nadavh@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/5607
Tested-by: Star_Automation <star@marvell.com>