cpufreq: add 'freq_table' in struct cpufreq_policy
freq table is not per CPU but per policy, so it makes more sense to
keep it within struct cpufreq_policy instead of a per-cpu variable.

This patch does it. Over that, there is no need to set policy->freq_table
to NULL in ->exit(), as policy structure is going to be freed soon.

Change-Id: Id3aca8d53f3d15d6311519b6f6ca5980619ee6d4
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Neta Zur <neta@marvell.com>

Conflicts:
	drivers/cpufreq/loongson2_cpufreq.c
	drivers/cpufreq/pasemi-cpufreq.c
	drivers/cpufreq/ppc-corenet-cpufreq.c
	drivers/cpufreq/sparc-us2e-cpufreq.c
	drivers/cpufreq/sparc-us3-cpufreq.c
	drivers/cpufreq/tegra-cpufreq.c