spi: a38x: Implement "50MHZ SPI AC timing" Erratum No. FE-9144572
Description:
The device SPI interface supports frequencies of up to 50 MHz.
However, due to this erratum, when the device core clock is 250 MHz
and the SPI interfaces is configured for 50MHz
SPI clock and CPOL=CPHA=1, there might occur data corruption on reads from the SPI device.

Workaround:
Work in one of the following configurations:
1. Set CPOL=CPHA=0 in "SPI Interface Configuration Register".
2. Set TMISO_SAMPLE value to 0x2 in "SPI Timing Parameters 1 Register" before setting the interface.

Change-Id: Ib77e4b66a5aaa5c049deb38ae3b049891e47512f
Signed-off-by: Nadav Haklai <nadavh@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/19626
Tested-by: Star_Automation <star@marvell.com>