defconfig: mvebu: add mergeable defconfig extra features
This defconfig is built to add extra features that are not covered in the
default mvebu defconfig.

For example cgroups, fhandle and other functionality.

This file is a starting point to make it easy for users to include more
features in the kernel build and will include more mini pcie based cards, usb
devices and others.

To use it, run -
./scripts/kconfig/merge_config.sh arch/arm/configs/mvebu_lsp_defconfig arch/arm/configs/mvebu_extra_defconfig

Signed-off-by: Rabeeh Khoury <rabeeh@solid-run.com>