vfio: xhci: add support for XHCI reset hook
The VFIO reset hook is called every time a platform device is
passed to a guest or removed from a guest.
When the XHCI device is unbound from the host, the host driver
disables the XHCI clock so when the device is passed to the
guest it becomes dis-functional.
This initial implementation uses the VFIO reset hook to enable
the XHCI clocks on behalf of the guest.

Change-Id: I1e729645b9cb600df39b5bd8ff213ec34a137d2d
Signed-off-by: Yehuda Yitschak <yehuday@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/29163
Tested-by: Star_Automation <star@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Reviewed-by: Shadi Ammouri <shadi@marvell.com>