fix: mmc: disable the command conflict error
The "command conflict error" feature is enabled by
default to detect command conflict between host and
device and report error. When the timing is very high
such as in SD SDR104 mode, the command conflict error may
be reported, which leads to error printing and device reset.
This patch disables this feature that do not to detect and
report the conflict, which does not affect SD/MMC
functionalities.

Change-Id: Ide40c6fb7c5a84e30061b93f82f5e39667580d98
Signed-off-by: Victor Gu <xigu@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/30856
Reviewed-by: Shadi Ammouri <shadi@marvell.com>
Tested-by: Star_Automation <star@marvell.com>
Reviewed-by: Wilson Ding <dingwei@marvell.com>