mmc: xenon: support voltage PAD setting for MMC PHY
Some version of XENON MMC PHY need external input for MMC
PHY voltage, either in 3.3v or 1.8v.
Normally it is set by SoC level PAD register out of MMC.
This patch adds support of PAD register setting.
The PAD register address need to be added as second register
range within SDHCI DT node.
An DT property "xenon,fixed-1-8v-pad-ctrl" is added to indicate
whether the PAD voltage setting will always be fixed to 1.8v,
which is used for eMMC and some SDIO devices with fixed 1.8v
signal voltage.
If this property is not set, it will be treated as SD with
initial 3.3v voltage and dynamic voltag switch.

Change-Id: I8f79571524c2f529110f48cd25fd54d5b2184c6d
Signed-off-by: Victor Gu <xigu@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/32165
Tested-by: Star_Automation <star@marvell.com>
Reviewed-by: Wilson Ding <dingwei@marvell.com>