mtd: pxa3xx_nand: improve clock handling and probe deferral
This patch improves style of clock handling. Also, when devm_clk_get()
returns -EPROBE_DEFER, the warning will not be printed.

result of devm_clk_get() takes possible probe deferral
into consideration, which is needed on a7k/a8k, because CP110 system
controller clocks are registered as normal driver (not using
CLK_OF_DECLARE macro for probing).

On some platforms, the clocks might be registered by a platform
driver. When this is the case, the clock platform driver may very well
be probed after NAND driver, in which case the first probe() invocation
of NAND driver will receive -EPROBE_DEFER as the return value of
devm_clk_get().

The current code handles that as a normal error, and simply assumes
that this means that the system doesn't have a clock for the this 
unit, and continues probing without calling clk_prepare_enable(). 

Unfortunately, this doesn't work with units in which the clock is
provided by another platform driver. In order to fix this situation,
we handle the -EPROBE_DEFER error condition specially, and abort the
driver probe(). It will be retried later automatically, the
clock will be available, devm_clk_get() will succeed, and the probe()
will continue with the clock prepared and enabled as expected.

In practice, such issue is seen on the ARM64 Marvell 7K/8K platform,
where the clocks are registered by a platform driver.

This patch is part of patch-set which adds
mainline mvebu clock gating support [18/21].

Change-Id: Ifcb9fdcd11cf7d38ba6493cab8b6574153ee8555
Signed-off-by: Marcin Wojtas <mw@semihalf.com>
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/32403
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>
Tested-by: Omri Itach <omrii@marvell.com>