dts: a38x: a39x: pl310-cache disable double-linefill
PL310 double-linefill is disabled in this commit, in order
to overcome an instability issue encountered under heavy system stress.

In continue to the deadlock that was generated while using ARM CA9 ACP
and L2 PL310, double linefill imposes a dependency between an
outstanding read and incoming write.
This dependency violates PCIE overtaking rule which creates a deadlock
with specific use cases.

Change-Id: Ia340b905051da85c0dacb9fb54dbfd9cf7d37d28
Signed-off-by: Yan Markman <ymarkman@marvell.com>
Signed-off-by: Igal Liberman <igall@marvell.com>
Signed-off-by: Nadav Haklai <nadavh@marvell.com>
Cc: stable@vger.kernel.org
Reviewed-on: http://vgitil04.il.marvell.com:8080/42874
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>