ahci: comphy: cp110: call comphy driver to configure the phy into SATA mode
Until now, in MVEBU AHCI Linux assumes that comphy (SerDes) was
configured by u-boot.
Now there's a new Linux comphy driver which configures the comphy (SerDes).

This patch adds a new flow in order to be able to use the new comphy
driver for cp110:
1. Power down AHCI macs
2. Configure the corresponding comphy (comphy driver).
3. Power up AHCI macs
4. Check if comphy PLL was locked

The new code is placed under "marvell,armada-cp110-ahci" because it's
required only for cp110 AHCI.

Change-Id: I3bd3b47fac0616f07ecf23dd8a779125d0a32abe
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/33629
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>
Tested-by: Star_Automation <star@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>