fix: arm64: defconfig: stop ignoring possible el1 async abort errors
This patch disables CONFIG_IGNORE_ASYNC_ABORT which was added
to ignore asynchronous external abort in EL1.

The EL1 asynchronous external abort happened when a PCIe switch
with no endpoint was connected to the PCIe slot.

Masking EL1 errors might hide other existing issues and it’s not
the correct way to handles such cases.

Change-Id: I03a8e6cfad1fae0f874b02fbd6990b549da60810
Signed-off-by: Omri Itach <omrii@marvell.com>
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/34724
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Wilson Ding <dingwei@marvell.com>