fix: dts: cp110: fix interrupt mapping for eip197
EIP197 engine was tested only on CP0.
The available interrupts mapping for CP1 is not correct.
This patch fixes the interrupt mapping, this allows to use EIP197
on CP1 too.

In addition, add eip197 to cp110-1 (disabled) in configuration
'A' and 'D'.
It's disabled because, currently, EIP197 can't run on both CPs
in the same time, so we add it as a place holder.

Change-Id: I40e25ee9e9e8da31cc2067696bf7a6e5a554a986
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/35377
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Ofer Heifetz <oferh@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>