fix: clk: cp110: Add mmc clock to cp-110 system controller
Add missing fixed clock to CP-110 system controller.
The MMC clock is dedicated to drive SDHCI interface
and its frequency is 400MHz (2/5 of PLL source clock)
The SDIO interface should be bound to this clock and
not the core clock as in the older code.
Wrong clock bounding results in SDHCI maximum frequency
shown as 250MHz instead of 400MHz

Change-Id: Ibef8392cca850f77644209a5f76e35ac84d04bb3
Signed-off-by: Konstantin Porotchkin <kostap@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/37359
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>