crypto: inside-secure: add resource sharing between US and KS
This patch adds new module param called "rings".
This parameters allows the user to tell the driver how many
rings are available for its use.

The driver expects to read the number of rings for each engine
in the following way:
insmod <name.ko> rings= <engine1 rings>, <engine2 rings>
For example: crypto_safexcel.ko rings = 4,0
	4 rings of the first engine to be used by the Kernel
	0 rings of the second engine to be used by the Kernel
If "rings" is not state, all rings will be used by the Kernel.

This patch allows to divide the HW rings between the Kernel and
the user-space. The Kernel will do basic initialization of all
rings, but won't use the rings which are dedicated to US.

Change-Id: I67a811dc495ea9f0d48244a879c7b1226d4fbec2
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/37583
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>