pinctrl: armada-37xx: Add pin controller support for Armada 37xx
The Armada 37xx SoC come with 2 pin controllers: one on the south
bridge (managing 28 pins) and one on the north bridge (managing 36 pins).

At the hardware level the controller configure the pins by group and not
pin by pin. This constraint is reflected in the design of the driver:
only the group related functions are implemented.

Cherry-picked from branch linux-4.12-devel and minor changes because of
the different pinctrl APIs between 4.12 and 4.4.52:
  pinctrl_utils_free_map -> pinctrl_utils_dt_free_map
  devm_pinctrl_register -> pinctrl_register

Change-Id: I49c8397abfa380f389dba074efbb9cadf5804068
Signed-off-by: Gregory CLEMENT <gregory.clement@free-electrons.com>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
Signed-off-by: Ken Ma <make@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/42769
Reviewed-by: Victor Gu <xigu@marvell.com>
Tested-by: Victor Gu <xigu@marvell.com>