pinctrl: armada-37xx: Add gpio support
GPIO management is pretty simple and is part of the same IP than the pin
controller for the Armada 37xx SoCs.  This patch adds the GPIO support to
the pinctrl-armada-37xx.c file, it also allows sharing common functions
between the gpiolib and the pinctrl drivers.

Cherry-picked from branch linux-4.12-devel and some changes because of
the different kernel gpio framework:
  1. In rc12 private pinctrl info is added to gpio chip as data member
     but 4.4.52 gpio chip has no such data member, so in gpio hook
     functions, private pinctrl info is not got from gpio chip data
     but got as a container of gpio chip;
  2. Since there is no gpio chip data member in 4.4.52, there is no
     API of devm_gpiochip_add_data, it is replaced by gpiochip_add.

Change-Id: I252b8646eab9180d15509695af7a245b7d4012e3
Signed-off-by: Gregory CLEMENT <gregory.clement@free-electrons.com>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
Signed-off-by: Ken Ma <make@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/42770
Reviewed-by: Victor Gu <xigu@marvell.com>
Tested-by: Victor Gu <xigu@marvell.com>