fix: usb: gadget: u3d: enable EP during PM complete stage
The Marvell USB3 device driver relies on the mass storage kernel
thread. Once the EP is enabled in USB3 device driver during system
resume, the USB host such as PC starts the transfer, and the mass storage
function thread is expected to handle the transfer.
If the transfer can not be finished within 5 second, the USB host will
set USB device to suspend state, in which state the USB3 device can not
work even after the mass storage kernel thread is resumed.
To avoid this case, the time gap should be as less as possible between
EP enabling and the mass storage thread is resumed.
This patch moves the mvc2_start(used to enable EP) to power management
.complete callback routine, which will be executed after all device are
resumed, and before starting to resume the kernel thread.
JIRA A3700-858

Change-Id: Id4b5d5abc763121fc384e4a8c0f307d68c3d3610
Signed-off-by: Victor Gu <xigu@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/38883
Reviewed-by: Omri Itach <omrii@marvell.com>