fix: usb: gadget: u3d: remove USB3 reset during disconnection
The routine mvc2_hw_reset() is used to do whole USB3 device controller
reset, which should be used after VBUS is on. When USB3 VBUS is on,
it means USB3 device is plugged into host, driver should prepare to
communicate with host, and the reset is mandatory.
If VBUS is off, it means device is disconnected from host and the 
USB3 device functionality stops, thus it is not necessary to do reset
during VBUS disconnection.
This patch removes the reset during USB3 device VBUS disconnection.
The USB3 device suspend and resume long run instability issue is fixed
after removing this reset during disconnection.
JIRA A3700-959.

Change-Id: Idb506c9392e35f3adf81d375c31a67c49358b6b9
Signed-off-by: Victor Gu <xigu@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/39254
Reviewed-by: Omri Itach <omrii@marvell.com>
Tested-by: iSoC Platform CI <ykjenk@marvell.com>