dts: cp110: comphy: add property of dfx-reg and sar-data
DFX register is needed to configure PCIe clock MUX
when clock direction is out and lane5 is indicated as
clock source. The property of dfx-reg indicates the DFX
register space information.

Sample At Reset(SAR) is needed to get PCIe clock direction
when COMPHY is initialized for PCIe. The property of
sar-data is the phandle to SAR device in DTS.

Change-Id: I4eb27dd7330c53b62a9d4d78f9866ffd9f230445
Signed-off-by: Evan Wang <xswang@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/40539
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>