fix: usb: ehci-orion: disable streaming for Armada 3700
Set stream disable to guarantee the DDR access in low bandwidth
system as Armada3700.
Set "stream disable" ensures that overruns/underruns of the
latency FIFO are eliminated for low bandwidth systems where
the RX and TX buffers are sufficient to contain the entire
packet. Enabling stream disable also has the effect of ensuring
the TX latency FIFO is filled to capacity before the packet is
launched onto the USB.

Change-Id: I04ad21618033e54c36164fe88b31ffbf5ae70c39
Signed-off-by: jinghua <xigu@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/41613
Tested-by: iSoC Platform CI <ykjenk@marvell.com>