fix: uart: a37x0: fix start tx flow
When Armada37x0 works with pulse interrupt, the TX interrupt
is triggered only when TX FIFO state transmites in following
two scenarios:
- FIFO FULL => FIFO NON-FULL
- FIFO EMPTY => FIFO NON-FULL

In order to transmit the sequential data, the driver send out
the first byte to toggle the FIFO state. And the rest bytes
are supposed to be sent out until TX READY interrupt comes.

Now, the driver enqueues the first byte to the TX FIFO prior
to enable the TX interrupt. At that time, the first byte has
already been shifted out. The pulse interrupt state is not
able to be retained. Thus, the following issue is observed
on UART1 port, which uses pulse interrupt.

When transmitting a set of sequential data, the data is not
received with the expected intervals.

original data stream:
ABCDE-<interval>-ABCDE-<interval>-ABCDE-...

observed data stream:
A-<interval>-BCDEABCDE-<interval>-A-<interval>-BCDE...

This patch corrected the sequence of sending the first byte
and enabling TX interrupt in mvebu_uart_start_tx().

Change-Id: I39b1673f23dd2cea417116bc602b800877a7111b
Signed-off-by: allen yan <yanwei@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/42382
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Victor Gu <xigu@marvell.com>
Reviewed-by: Wilson Ding <dingwei@marvell.com>