fix: a8k: pcie: add pcie prelink configurations
The PCIe prelink configurations are implemented in uboot
currently, and in order to remove the dependency
on bootloader, it is necessary to plant such configurations
to kernel.

The prelink confgiurations are only needed when PCIe
initialization sequence is completely done in kernel
driver, including COMPHY power up. The patch take the
COMPHY count as the precondition to run PCIe prelink
configurations.

Change-Id: I4085aff1e57e2beafc4c5517d7fedfaf1a0f9d7a
Signed-off-by: Evan Wang <xswang@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/44163
Reviewed-by: Omri Itach <omrii@marvell.com>
Tested-by: Hanna Hawa <hannah@marvell.com>