fix: dts: a3900: remove pcie comphy specification for avoid network error
PCIe comphy driver issues: The PPv22 mac0 and mac2 would stop functioning,
                           when specify the PCIEx2 in the DTS.

This patch is a WA for this issue, which mvppv2 will stop functioning
with the PCIEx2 setting in DTS. Remove the pcie comphy specification
and depend on uboot PCIe initialization, the network and PCIE are works.

Please remove this patch after fix the PCIe comphy driver issues.

Change-Id: I7cb1606b251482f1bbf5124805c74fcb7c125b6d
Signed-off-by: yachun <yachun@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/44421
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Nadav Haklai <nadavh@marvell.com>