fix: clk: mvebu: a37xx: correct the load level calculation
There is an issue when calculating load level that the
load level width is 2 bits, so its valid value calculation
shoudl be (level & 0x3) not (level * 0x3).

The issue will cause DVFS set scaling frequency does not
work because there is possiblity that the target CPU load
level exceeds limitation.

The patch correct the typo issue for calculation of load
level.

Change-Id: Ida29b3894a79b50491ceefef032be1e5ce5c489c
Signed-off-by: Evan Wang <xswang@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/44327
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Victor Gu <xigu@marvell.com>