fix: pinctrl: armada-37xx: ack pending GPIO irq
The GPIO polarity registers are saved during system suspend and
restored during resume. If there is pending GPIO irq during the
resume, the GPIO interrupt will be triggered when re-config the
GPIO interrupt enable and polarity registers.
That time the system infrastructure modules are already initialized
including GICv3, but the IO drivers are still not resumed, it
will lead to unexpected system stuck if gets IO interrupt that time.
In order to fix this issue, need to ack the pending GPIO irq
during resume before restoring the enable and polarity setting.

Change-Id: I2dde3b1492f1d45d3430b695dda65bcc51afb264
Signed-off-by: Victor Gu <xigu@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/44381
Reviewed-by: Marcin Wojtas <mw@semihalf.com>