fix: a8k: pcie: update sequence of PCIe power down/up
In PCIe test with different EP devices, it is found that
the GenX or link width detected false after PCIe link up
for some EP devices, for example, the speed is detected
as Gen1 for Gen2 EP device, etc..
According to design, it is necessary to put PCIe relevant
modules to initial state, for example PCIe MAC reset, EP
reset, etc..

So the PCIe power down/up sequence is updated for Linux
boot with design suggestions:
1)	PCIe - Power Down
	a. PCIe MAC - Soft Reset
	b. ComPhy Hard Reset
2)	EP Reset (One reset signal to all CPx PCIe EP card)
	a. EP Reset
	b. Delay 1ms
	c. EP Reset de-asserted
3)	PCIe - Power Up
	a. PCIe MAC - Soft Reset de-asserted
	b. ComPhy init
	c. Configure PCIE MAC.
	d. LTSSM_En
	e. Check Link with time out 1 Sec
	f. wait 300ms for MAC negotiation done.

The steps 1).a, 1).b, 3).a and 3).b in the sequence are
implemented in CP110 COMPHY driver in another patch.
Other steps are implemented in this patch.

The patch and previous pacthes implemented the whole PCIe
power down/up sequence, and fixed the issue that GenX or
link width detected false for some EP device, such as SATA
over PCIe device.

Change-Id: Idb760ea392971047692e40435542cf7f0187fe8a
Signed-off-by: Evan Wang <xswang@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/44417
Reviewed-by: Victor Gu <xigu@marvell.com>
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>