fix: ap806: thermal: fix invalid handling of private struct
AP806 thermal sensor registers as a sensor of a Device
Tree thermal zone, it's binded differently from other
thermal sensors supported by this driver, such as
armada-370/380/XP and CP110.
When sensor initialization, the required information are
stored in private struct. For SoC except AP806 the private
struct is stored in thermal struct when register, but for
AP806 it is stored in an internal struct of thermal zone.
So AP806 can not get correct private data via thermal
struct when resume, which causes callback pointer to sensor
initialization is invalid and then causes CPU stuck.

To fix the issue, this patch registers the private thermal
struct as driver data instead of just the thermal struct,
which only changed the place to store private thermal and
for other SoCs the private thermal are same. So all SoCs
including AP806 can get its own private data correctly
which is required by sensor initialization when resume.

Change-Id: I5948c7220da69a6b7bd84ec97a7c51af001843ed
Signed-off-by: Evan Wang <xswang@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/44492
Reviewed-by: Omri Itach <omrii@marvell.com>
Tested-by: Omri Itach <omrii@marvell.com>