dts: cp110: differentiate SAR with new compatible string
On a8k, there are CP-0 and CP-1, and their functions are
not completely same in sample at reset, so it is necessary
to differentiate them.
The patch differentiates CP0 SAR and CP1 SAR with different
compatible string.
For A7K, the string is "marvell,a70x0-sample-at-reset".
For A8K CP0, it is "marvell,a80x0-cp0-sample-at-reset".
For A8K Cp1, it is "marvell,a80x0-cp1-sample-at-reset".

Change-Id: I29cc45fc39b781254b06afa5a36a0e51f32f2595
Signed-off-by: Evan Wang <xswang@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/44671
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>