comphy: cp110: add PCIe MAC reset when power off and on
According to PCIe initialization sequence as below:
1) PCIe - Power Down
   a. PCIe MAC - Soft Reset
   b. ComPhy Hard Reset
2) EP Reset
   a. EP Reset
   b. Delay 1ms
   c. EP Reset de-asserted
3) PCIe - Power Up
   a. PCIe MAC(Soft Reset de-asserted)
   b. ComPhy init
   c. Configure PCIE MAC.
   d. LTSSM_En
   e. Check Link with time out 1 Sec
   f. wait 300ms for MAC negotiation done.

As part of PCIe initialization sequency, the PCIe MAC reset
should be implemented before PCIe hard reset, and release
reset before PCIe comphy initialization.
According to SoC design, PCIe port0 is connected to COMPHY
lane0, and port1 is connected to lane4 and port2 is
connected to lane5.

The patch implements the reset as below:
1) When COMPHY power off, the PCIe MAC will reset before
   COMPHY hard reset via resetting corresponding COMPHY
   lanes.
2) At the beginning COMPHY power on, the PCIe reset release
   is done before PIPE selector set via configuring
   corresponding COMPHY lanes.

The patch implemented part of PCIe power up sequence and
not involve function changes.

Change-Id: I73a843e20db220ec91bd0bd6f08377edd450f232
Signed-off-by: Evan Wang <xswang@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/44892
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>